Make service scoping consistent between SQLite and SQL Server providers
This change makes the SqliteTypeMapper Singleton so that the type mappings are cached, which is consistent with what SQL Server does.

Also, ISqlServerSqlGenerator is made into a server that can be overidden by providers if needed.

See Issue #2168
