+++ src/EntityFramework.Relational/EntityFramework.Relational.csproj
@@ -71,6 +71,7 @@
     <Compile Include="IRelationalTypeMapper.cs" />
     <Compile Include="IRelationalValueBufferFactoryFactory.cs" />
     <Compile Include="ISqlGenerator.cs" />
+    <Compile Include="ISqlStatementExecutor.cs" />
     <Compile Include="Metadata\IRelationalMetadataExtensionProvider.cs" />
     <Compile Include="Metadata\RelationalCollectionReferenceBuilderExtensions.cs" />
     <Compile Include="Metadata\RelationalReferenceReferenceBuilderExtensions.cs" />
