+++ src/EntityFramework.Relational/RelationalEntityServicesBuilderExtensions.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Data.Entity.Relational
                 .AddSingleton<IParameterNameGeneratorFactory, ParameterNameGeneratorFactory>()
                 .AddSingleton<IComparer<ModificationCommand>, ModificationCommandComparer>()
                 .AddSingleton<IMigrationIdGenerator, MigrationIdGenerator>()
-                .AddSingleton<ISqlStatementExecutor, SqlStatementExecutor>()
+                .AddSingleton<SqlStatementExecutor>()
                 .AddSingleton<UntypedValueBufferFactoryFactory>()
                 .AddSingleton<TypedValueBufferFactoryFactory>()
                 .AddSingleton<IMigrationModelFactory, MigrationModelFactory>()
@@ -41,6 +41,7 @@ namespace Microsoft.Data.Entity.Relational
                 .AddScoped<RelationalDatabaseFactory>()
                 .AddScoped<RelationalValueGeneratorSelector>()
                 .AddScoped<CommandBatchPreparer>()
+                .AddScoped(p => GetStoreServices(p).SqlStatementExecutor)
                 .AddScoped(p => GetStoreServices(p).ModelDiffer)
                 .AddScoped(p => GetStoreServices(p).HistoryRepository)
                 .AddScoped(p => GetStoreServices(p).MigrationSqlGenerator)
