+++ src/EntityFramework.SqlServer/Extensions/SqlServerEntityServicesBuilderExtensions.cs
@@ -27,12 +27,12 @@ namespace Microsoft.Framework.DependencyInjection
                 .TryAdd(new ServiceCollection()
                     .AddSingleton<SqlServerModelBuilderFactory>()
                     .AddSingleton<ISqlServerValueGeneratorCache, SqlServerValueGeneratorCache>()
-                    .AddSingleton<ISqlServerSequenceValueGeneratorFactory, SqlServerSequenceValueGeneratorFactory>()
                     .AddSingleton<ISqlServerSqlGenerator, SqlServerSqlGenerator>()
                     .AddSingleton<SqlServerTypeMapper>()
-                    .AddSingleton<SqlServerModificationCommandBatchFactory>()
                     .AddSingleton<SqlServerModelSource>()
                     .AddSingleton<SqlServerMetadataExtensionProvider>()
+                    .AddScoped<ISqlServerSequenceValueGeneratorFactory, SqlServerSequenceValueGeneratorFactory>()
+                    .AddScoped<SqlServerModificationCommandBatchFactory>()
                     .AddScoped<SqlServerValueGeneratorSelector>()
                     .AddScoped<SqlServerDataStoreServices>()
                     .AddScoped<SqlServerDataStore>()
