+++ src/EntityFramework.Sqlite/SqliteEntityFrameworkServicesBuilderExtensions.cs
@@ -26,9 +26,9 @@ namespace Microsoft.Framework.DependencyInjection
                     .AddSingleton<SqliteValueGeneratorCache>()
                     .AddSingleton<SqliteSqlGenerator>()
                     .AddSingleton<SqliteMetadataExtensionProvider>()
-                    .AddScoped<SqliteTypeMapper>()
-                    .AddScoped<SqliteModificationCommandBatchFactory>()
+                    .AddSingleton<SqliteTypeMapper>()
                     .AddSingleton<SqliteModelSource>()
+                    .AddScoped<SqliteModificationCommandBatchFactory>()
                     .AddScoped<SqliteDataStoreServices>()
                     .AddScoped<SqliteDataStore>()
                     .AddScoped<SqliteDataStoreConnection>()
