+++ test/EntityFramework.SqlServer.Tests/SqlServerEntityFrameworkServicesBuilderExtensionsTest.cs
@@ -33,15 +33,15 @@ namespace Microsoft.Data.Entity.SqlServer.Tests
             // SQL Server dingletones
             VerifySingleton<SqlServerModelBuilderFactory>();
             VerifySingleton<ISqlServerValueGeneratorCache>();
-            VerifySingleton<ISqlServerSequenceValueGeneratorFactory>();
             VerifySingleton<ISqlServerSqlGenerator>();
             VerifySingleton<ISqlStatementExecutor>();
             VerifySingleton<SqlServerTypeMapper>();
-            VerifySingleton<SqlServerModificationCommandBatchFactory>();
             VerifySingleton<SqlServerModelSource>();
             VerifySingleton<SqlServerMetadataExtensionProvider>();
 
             // SQL Server scoped
+            VerifyScoped<SqlServerModificationCommandBatchFactory>();
+            VerifyScoped<ISqlServerSequenceValueGeneratorFactory>();
             VerifyScoped<SqlServerValueGeneratorSelector>();
             VerifyScoped<SqlServerDataStoreServices>();
             VerifyScoped<SqlServerDataStore>();
